Minor modifications
Changed the plotting function to give an option to draw both methods at the same time or just the first one. Changed it so that it plots the square S defined which surround the region K subject to the condition. Also added an option to save the figure to a pdf file in the same folder.